number guessing game
This is a number guessing game. It first asks the player for their name. Then it chooses a random number between 1 and 20 and gives the user three chances to guess the number. If the number that the player guesses is greater than or equal to the correct number minus three and also less than the correct number, the game will tell them to guess a little higher. If the player's guess is less than the correct number minus three, the game will tell the player that their guess is too low. If the player's guess is greater than the correct number, and less than the correct number plus three, the game will tell the player to guess a little lower. If the player's guess is greater than the correct number plus 3, the game will tell them that their guess is too high. The game will do this for the player's first two guesses. If the player guesses the correct number, the game will tell them that they guessed the correct number and the program will end. If the player does not guess the correct number in three turns, the game will tell them that they did not guess the right number and that they have used up all three strikes and the program will end.